refactor: Separate views and logics for writing story
