+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/DaySlide/ListFooterContainer.java
@@ -29,13 +29,7 @@ public class ListFooterContainer {
     private final Button addButtonInactivate;
 
     // 스토리 작성
-    private final ConstraintLayout storyWrapper;
-    private final EditText storyTitleEditText;
-    private final EditText storyContentEditText;
-    private final TextView completeTimeText;
-    private final TextView storyContentLengthText;
-    private final TextView aiButtonHelpText;
-    private final Button storyAiButton;
+    private final StoryContainer storyContainer;
 
     // 감정 선택
     private final ConstraintLayout emotionWrapper;
@@ -51,8 +45,6 @@ public class ListFooterContainer {
     private final MutableLiveData<Boolean> bottomButtonState = new MutableLiveData<>(false);
 
     private final int MOMENT_MAX_LENGTH = 1000;
-    private final int STORY_TITLE_MAX_LENGTH = 100;
-    private final int STORY_CONTENT_MAX_LENGTH = 1000;
 
     public ListFooterContainer(@NonNull View view) {
         // 모먼트 작성
@@ -67,13 +59,7 @@ public class ListFooterContainer {
         addButtonInactivate = view.findViewById(R.id.add_button_inactivate);
 
         // 스토리 작성
-        storyWrapper = view.findViewById(R.id.story_wrapper);
-        storyTitleEditText = view.findViewById(R.id.storyTitleEditText);
-        storyContentEditText = view.findViewById(R.id.storyContentEditText);
-        completeTimeText = view.findViewById(R.id.completeTimeText);
-        storyContentLengthText = view.findViewById(R.id.storyContentLengthText);
-        aiButtonHelpText = view.findViewById(R.id.aiButtonHelpText);
-        storyAiButton = view.findViewById(R.id.storyAiButton);
+        storyContainer = new StoryContainer(view.findViewById(R.id.story_wrapper));
 
         // 감정 선택
         emotionWrapper = view.findViewById(R.id.emotion_wrapper);
@@ -126,67 +112,9 @@ public class ListFooterContainer {
         addButtonInactivate.setOnClickListener(v -> {
         });
 
-        // storyTitleEditText 입력 감지
-        storyTitleEditText.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                if (s.length() > STORY_TITLE_MAX_LENGTH) {
-                    // 글자수 제한 초과
-                    storyTitleEditText.setText(s.subSequence(0, STORY_TITLE_MAX_LENGTH));
-                    storyTitleEditText.setTextColor(
-                        ContextCompat.getColor(view.getContext(), R.color.red));
-                    storyTitleEditText.requestFocus();
-                    storyTitleEditText.setSelection(STORY_TITLE_MAX_LENGTH);
-                } else {
-                    storyTitleEditText.setTextColor(
-                        ContextCompat.getColor(view.getContext(), R.color.black));
-                }
-            }
-        });
-
-        // storyContentEditText 입력 감지
-        storyContentEditText.addTextChangedListener(new TextWatcher() {
-            @Override
-            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start, int before, int count) {
-            }
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                setStoryContentLengthText(s.length());
-
-                if (s.length() > STORY_CONTENT_MAX_LENGTH) {
-                    // 글자수 제한 초과
-                    storyContentEditText.setText(s.subSequence(0, STORY_CONTENT_MAX_LENGTH));
-                    storyContentEditText.setTextColor(
-                        ContextCompat.getColor(view.getContext(), R.color.red));
-                    storyContentEditText.requestFocus();
-                    storyContentEditText.setSelection(STORY_CONTENT_MAX_LENGTH);
-                } else {
-                    storyContentEditText.setTextColor(
-                        ContextCompat.getColor(view.getContext(), R.color.black));
-                }
-            }
-        });
-
         // 감정 선택 감지
         emotionGridContainer.setSelectedEmotionObserver((Integer emotion) -> {
-            if (emotion > -1) {
-                setBottomButtonState(true);
-            } else {
-                setBottomButtonState(false);
-            }
+            setBottomButtonState(emotion > -1);
         });
 
         // 태그 개수 제한 감지
@@ -216,8 +144,7 @@ public class ListFooterContainer {
     }
 
     public void freezeStoryEditText() {
-        storyTitleEditText.setEnabled(false);
-        storyContentEditText.setEnabled(false);
+        storyContainer.freeze();
     }
 
     public void freezeEmotionSelector() {
@@ -270,16 +197,13 @@ public class ListFooterContainer {
         momentLengthText.setVisibility(View.GONE);
         submitButtonInactivate.setVisibility(View.GONE);
 
-        storyWrapper.setVisibility(View.VISIBLE);
-        completeTimeText.setText(completeTime);
+        storyContainer.setUiWritingStory(completeTime);
 
         setBottomButtonState(true);
     }
 
     public void setUiSelectingEmotion() {
-        storyContentLengthText.setVisibility(View.GONE);
-        aiButtonHelpText.setVisibility(View.GONE);
-        storyAiButton.setVisibility(View.GONE);
+        storyContainer.setUiCompleteStory();
 
         emotionWrapper.setVisibility(View.VISIBLE);
 
@@ -293,10 +217,5 @@ public class ListFooterContainer {
     private void setMomentLengthText(int count) {
         momentLengthText.setText(
             String.format(Locale.getDefault(), "%d / %d", count, MOMENT_MAX_LENGTH));
-    }
-
-    private void setStoryContentLengthText(int count) {
-        storyContentLengthText.setText(
-            String.format(Locale.getDefault(), "%d / %d", count, STORY_CONTENT_MAX_LENGTH));
     }
 }
