+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/DaySlide/StoryContainer.java
@@ -1,0 +1,111 @@
+package snu.swpp.moment.ui.main_writeview.DaySlide;
+
+import android.text.Editable;
+import android.text.TextWatcher;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.TextView;
+import androidx.constraintlayout.widget.ConstraintLayout;
+import androidx.core.content.ContextCompat;
+import java.util.Locale;
+import snu.swpp.moment.R;
+
+public class StoryContainer {
+
+    private final ConstraintLayout storyWrapper;
+    private final EditText storyTitleEditText;
+    private final EditText storyContentEditText;
+    private final TextView completeTimeText;
+    private final TextView storyContentLengthText;
+    private final TextView aiButtonHelpText;
+    private final Button storyAiButton;
+
+    private final int STORY_TITLE_MAX_LENGTH = 100;
+    private final int STORY_CONTENT_MAX_LENGTH = 1000;
+
+    public StoryContainer(View view) {
+        storyWrapper = (ConstraintLayout) view;
+        storyTitleEditText = view.findViewById(R.id.storyTitleEditText);
+        storyContentEditText = view.findViewById(R.id.storyContentEditText);
+        completeTimeText = view.findViewById(R.id.completeTimeText);
+        storyContentLengthText = view.findViewById(R.id.storyContentLengthText);
+        aiButtonHelpText = view.findViewById(R.id.aiButtonHelpText);
+        storyAiButton = view.findViewById(R.id.storyAiButton);
+
+        // storyTitleEditText 입력 감지
+        storyTitleEditText.addTextChangedListener(new TextWatcher() {
+            @Override
+            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
+            }
+
+            @Override
+            public void onTextChanged(CharSequence s, int start, int before, int count) {
+            }
+
+            @Override
+            public void afterTextChanged(Editable s) {
+                if (s.length() > STORY_TITLE_MAX_LENGTH) {
+                    // 글자수 제한 초과
+                    storyTitleEditText.setText(s.subSequence(0, STORY_TITLE_MAX_LENGTH));
+                    storyTitleEditText.setTextColor(
+                        ContextCompat.getColor(view.getContext(), R.color.red));
+                    storyTitleEditText.requestFocus();
+                    storyTitleEditText.setSelection(STORY_TITLE_MAX_LENGTH);
+                } else {
+                    storyTitleEditText.setTextColor(
+                        ContextCompat.getColor(view.getContext(), R.color.black));
+                }
+            }
+        });
+
+        // storyContentEditText 입력 감지
+        storyContentEditText.addTextChangedListener(new TextWatcher() {
+            @Override
+            public void beforeTextChanged(CharSequence s, int start, int count, int after) {
+            }
+
+            @Override
+            public void onTextChanged(CharSequence s, int start, int before, int count) {
+            }
+
+            @Override
+            public void afterTextChanged(Editable s) {
+                setStoryContentLengthText(s.length());
+
+                if (s.length() > STORY_CONTENT_MAX_LENGTH) {
+                    // 글자수 제한 초과
+                    storyContentEditText.setText(s.subSequence(0, STORY_CONTENT_MAX_LENGTH));
+                    storyContentEditText.setTextColor(
+                        ContextCompat.getColor(view.getContext(), R.color.red));
+                    storyContentEditText.requestFocus();
+                    storyContentEditText.setSelection(STORY_CONTENT_MAX_LENGTH);
+                } else {
+                    storyContentEditText.setTextColor(
+                        ContextCompat.getColor(view.getContext(), R.color.black));
+                }
+            }
+        });
+    }
+
+    public void freeze() {
+        storyTitleEditText.setEnabled(false);
+        storyContentEditText.setEnabled(false);
+    }
+
+    public void setUiWritingStory(String completeTime) {
+        storyWrapper.setVisibility(View.VISIBLE);
+        completeTimeText.setText(completeTime);
+    }
+
+    public void setUiCompleteStory() {
+        storyContentLengthText.setVisibility(View.GONE);
+        aiButtonHelpText.setVisibility(View.GONE);
+        storyAiButton.setVisibility(View.GONE);
+    }
+
+    private void setStoryContentLengthText(int count) {
+        storyContentLengthText.setText(
+            String.format(Locale.getDefault(), "%d / %d", count, STORY_CONTENT_MAX_LENGTH));
+    }
+}
